swarm/network: Extract peer and addr types out of protocol file
